backend/Automated_Sys_tmp/automation_system/requirements.txt
django-cors-headers
djangorestframework
Pillow
djangorestframework-simplejwt
django_ratelimit
drf-yasg
weasyprint